model-server/requirements.txt
fastapi[all]==0.115.6
redis==5.2.1
requests
onnxruntime
uvicorn
pillow

